Introduce "To Iobroker" device classes
Logic to map Matter devi es for the current types into ioBroker direction.

ALso includes Lock (which is yet missing for the other direction. Will get in Readme next.
ALso includes a "UtilityOnly" type which for now uses a socket device as basis, needs to be discussed or added on iobroker side.